--- components/common/CharmEditor/components/tweet/twitterJSUrl.ts
@@ -1,1 +1,0 @@
-export const twitterWidgetJs = 'https://platform.twitter.com/widgets.js';
